Refatorado os models e excluído da pasta classes (que também será excluída)
