--- src/classes/contaPoupanca.ts
@@ -1,8 +1,0 @@
-import { Conta } from './conta';
-import { Cliente } from './cliente';
-
-export class ContaPoupanca extends Conta {
-    constructor(numero: number, cliente: Cliente) {
-        super(numero, cliente);
-    }
-}
