+++ src/models/conta.model.ts
@@ -1,11 +1,36 @@
-import { Module } from '@nestjs/common';
-import { ContasService } from '../services/conta.service';
-import { ContasController } from './conta.controller';
-import { ClientesModule } from './cliente.model';
+export abstract class Conta {
+  protected saldo: number = 0;
 
-@Module({
-  imports: [ClientesModule],
-  providers: [ContasService],
-  controllers: [ContasController]
-})
-export class ContasModule { }
+  constructor(
+      public numero: number,
+      public cliente: Cliente
+  ) { }
+
+  depositar(valor: number): void {
+      this.saldo += valor;
+      console.log(`Depositado: R$${valor}. Saldo atual: R$${this.saldo}`);
+  }
+
+  sacar(valor: number): void {
+      if (this.saldo >= valor) {
+          this.saldo -= valor;
+          console.log(`Sacado: R$${valor}. Saldo atual: R$${this.saldo}`);
+      } else {
+          console.log('Saldo insuficiente.');
+      }
+  }
+
+  transferir(valor: number, contaDestino: Conta): void {
+      if (this.saldo >= valor) {
+          this.saldo -= valor;
+          contaDestino.depositar(valor);
+          console.log(`Transferido: R$${valor} para conta ${contaDestino.numero}. Saldo atual: R$${this.saldo}`);
+      } else {
+          console.log('Saldo insuficiente.');
+      }
+  }
+
+  consultarSaldo(): number {
+      return this.saldo;
+  }
+}
