+++ src/models/contaPoupanca.model.ts
@@ -1,0 +1,4 @@
+import { Conta } from './conta.model';
+import { Cliente } from './cliente.model';
+
+export class ContaPoupanca extends Conta { }
